+++ src/game/image.h
@@ -11,10 +11,10 @@
 struct Image
 {
   std::vector<u8> data{};
-  uvec2 dimensions{};
+  vec2 dimensions{};
 };
 
-Image image_init(const u8* data, const uvec2& dimensions);
+Image image_init(const u8* data, const vec2& dimensions);
 std::expected<Image, std::string_view> image_from_file(const std::filesystem::path& path);
 
 #endif
